third_party/blink/web_tests/external/wpt/css/css-nesting/parsing-expected.txt

This is a testharness.js-based test.
PASS .foo {
  & { color: green; }
}
PASS .foo {
  &.bar { color: green; }
}
PASS .foo {
  & .bar { color: green; }
}
PASS .foo {
  & > .bar { color: green; }
}
FAIL .foo {
  > & .bar { color: green; }
} assert_equals: expected "& > & .bar" but got "> & .bar"
FAIL .foo {
  > .bar { color: green; }
} assert_equals: expected "& > .bar" but got "> .bar"
FAIL .foo {
  + .bar & { color: green; }
} assert_equals: expected "& + .bar &" but got "+ .bar &"
PASS .foo {
  .test > & .bar { color: green; }
}
FAIL .foo {
  + .bar, .foo, > .lol { color: green; }
} assert_equals: expected "& + .bar, .foo, & > .lol" but got "+ .bar, .foo, > .lol"
PASS .foo {
  &:is(.bar, &.baz) { color: green; }
}
PASS .foo {
  .bar& { color: green; }
}
PASS .foo {
  .bar > & { color: green; }
}
PASS .foo, .bar {
  & + .baz, &.qux { color: green; }
}
PASS .foo {
  & .bar & .baz & .qux { color: green; }
}
PASS .foo {
  @media (min-width: 50px) { color: green; }
}
PASS main {
  & > section, & > article {
  & > header { color: green; }
}
}
Harness: the test ran to completion.
